+++ packages/extensions/core/src/lib/plugins/semantics-validation/validators/paths-validator.ts
@@ -14,7 +14,6 @@ const operationKeys = new Set(["get", "post", "put", "delete", "options", "head"
  */
 export function validatePaths(spec: oai3.Model, resolve?: ResolveReferenceFn): SemanticError[] {
   const resolveReference = createReferenceResolver(spec, resolve);
-
   const paths = spec.paths;
   const errors: SemanticError[] = [];
   for (const [uri, pathItem] of Object.entries(paths)) {
@@ -66,6 +65,7 @@ function findPathParameter(
 
 function createPathParameterEmptyError(uri: string): SemanticError {
   return {
+    level: "error",
     code: SemanticErrorCodes.PathParameterEmtpy,
     message: `A path parameter in uri ${uri} is empty.`,
     params: { uri },
@@ -76,6 +76,7 @@ function createPathParameterEmptyError(uri: string): SemanticError {
 function createPathParameterMissingDefinitionError(uri: string, paramName: string, methods: string[]): SemanticError {
   const missingStr = methods.map((str) => `'${str}'`).join(", ");
   return {
+    level: "error",
     code: SemanticErrorCodes.PathParameterMissingDefinition,
     message: `Path parameter '${paramName}' referenced in path '${uri}' needs to be defined in every operation at either the path or operation level. (Missing in ${missingStr})`,
     params: { paramName, uri, methods },