+++ packages/libs/codemodel/src/model/common/code-model.ts
@@ -2,22 +2,10 @@ import { Metadata } from "./metadata";
 import { Schemas } from "./schemas";
 import { Info } from "./info";
 import { OperationGroup } from "./operation";
-import { DeepPartial, enableSourceTracking, Initializer } from "@azure-tools/codegen";
+import { DeepPartial, enableSourceTracking } from "@azure-tools/codegen";
 import { Parameter } from "./parameter";
 import { ValueOrFactory, realize, sort } from "@azure-tools/linq";
-
-/** The security information for the API surface */
-export interface Security {
-  /** indicates that the API surface requires authentication */
-  authenticationRequired: boolean;
-}
-
-export class Security extends Initializer implements Security {
-  constructor(public authenticationRequired: boolean, objectInitializer?: DeepPartial<Security>) {
-    super();
-    this.apply(objectInitializer);
-  }
-}
+import { Security } from "./security";
 
 /** the model that contains all the information required to generate a service api */
 export interface CodeModel extends Metadata {