+++ packages/libs/configuration/src/configuration-manager/configuration-file.ts
@@ -1,7 +1,12 @@
 import { AutorestLogger, CodeBlock, OperationAbortedException, parseCodeBlocks } from "@autorest/common";
 import { DataHandle, DataSink } from "@azure-tools/datastore";
-import { ParentFolderUri, IsUri } from "@azure-tools/uri";
-import { AutorestRawConfiguration } from "../autorest-raw-configuration";
+import { parentFolderUri, isUri } from "@azure-tools/uri";
+import { AutorestNormalizedConfiguration } from "../autorest-normalized-configuration";
+import {
+  autorestConfigurationProcessor,
+  AUTOREST_CONFIGURATION_SCHEMA,
+  RawConfiguration,
+} from "../configuration-schema";
 import { desugarRawConfig } from "../desugar";
 import { arrayOf } from "../utils";
 
@@ -21,7 +26,7 @@ export interface ConfigurationFile {
 
 export interface ConditionalConfiguration {
   condition?: string;
-  config: AutorestRawConfiguration;
+  config: AutorestNormalizedConfiguration;
 }
 
 export const readConfigurationFile = async (
@@ -29,7 +34,7 @@ export const readConfigurationFile = async (
   logger: AutorestLogger,
   sink: DataSink,
 ): Promise<ConfigurationFile> => {
-  const parentFolder = ParentFolderUri(configFile.originalFullPath);
+  const parentFolder = parentFolderUri(configFile.originalFullPath);
   const base: ConfigurationFile = {
     type: "file",
     fullPath: configFile.originalFullPath,
@@ -38,7 +43,7 @@ export const readConfigurationFile = async (
 
   // load config
   const hConfig = await parseCodeBlocks(logger, configFile, sink);
-  if (hConfig.length === 1 && hConfig[0].info === null && configFile.Description.toLowerCase().endsWith(".md")) {
+  if (hConfig.length === 1 && hConfig[0].info === null && configFile.description.toLowerCase().endsWith(".md")) {
     // this is a whole file, and it's a markdown file.
     return { ...base };
   }
@@ -56,12 +61,12 @@ const codeBlockToRawConfig = async (
   parentFolder: string | null,
   codeBlock: CodeBlock,
 ): Promise<ConditionalConfiguration> => {
-  const config = await codeBlock.data.ReadObject<AutorestRawConfiguration>();
-  if (!config) {
+  const rawConfig = await codeBlock.data.readObject<RawConfiguration<typeof AUTOREST_CONFIGURATION_SCHEMA>>();
+  if (!rawConfig) {
     return { config: {} };
   }
 
-  if (typeof config !== "object") {
+  if (typeof rawConfig !== "object") {
     logger.trackError({
       code: "",
       message: "Syntax error: Invalid YAML object.",
@@ -70,6 +75,19 @@ const codeBlockToRawConfig = async (
     throw new OperationAbortedException();
   }
 
+  const result = autorestConfigurationProcessor.processConfiguration(rawConfig, { logger });
+  if ("errors" in result) {
+    for (const error of result.errors) {
+      logger.trackError({
+        code: error.code,
+        message: error.message,
+        source: [{ document: codeBlock.data.key, position: { path: error.path } }],
+      });
+    }
+    throw new OperationAbortedException();
+  }
+
+  const config: AutorestNormalizedConfiguration = result.value;
   // for ['input-file','try-require', 'require'] paths, we're going to create a node that contains
   // a map of the path to the folder from which the configuration file
   // that loaded it was specified.
@@ -77,11 +95,11 @@ const codeBlockToRawConfig = async (
   // this will enable us to try to load relative paths relative to the folder from which it was read
   // rather than have to rely on the pseudo $(this-folder) macro (which requires updating the file)
   config.__parents = {};
-  const kinds: Array<keyof AutorestRawConfiguration> = ["input-file", "require", "try-require", "exclude-file"];
+  const kinds: Array<keyof AutorestNormalizedConfiguration> = ["input-file", "require", "try-require", "exclude-file"];
   for (const kind of kinds) {
     if (config[kind]) {
       for (const location of arrayOf<string>(config[kind])) {
-        if (!IsUri(location)) {
+        if (!isUri(location)) {
           config.__parents[location] = parentFolder;
         }
       }