+++ packages/libs/configuration/src/configuration-manager/configuration-manager.test.ts
@@ -167,12 +167,12 @@ describe("ConfigurationManager", () => {
         configs: [
           {
             config: {
-              directive: { from: "swagger-document", transform: "$.swagger=true" },
+              directive: [{ from: "swagger-document", transform: "$.swagger=true" }],
             },
           },
           {
             config: {
-              directive: { from: "openapi-document", transform: "$.openapi3=true" },
+              directive: [{ from: "openapi-document", transform: "$.openapi3=true" }],
             },
           },
         ],
@@ -243,7 +243,28 @@ describe("ConfigurationManager", () => {
       expect(output["namespace"]).toEqual("FooBarOverride.Client");
     });
 
-    it("interpolate from the last block", async () => {
+    it("interpolate from previous block", async () => {
+      manager.addConfigFile({
+        type: "file",
+        fullPath: "/dev/path/readme.md",
+        configs: [
+          {
+            config: {
+              name: "FooBar",
+            },
+          },
+          {
+            config: {
+              namespace: "$(name).Client",
+            },
+          },
+        ],
+      });
+      const output = await manager.resolveConfig();
+      expect(output["namespace"]).toEqual("FooBar.Client");
+    });
+
+    it("interpolate with value from same block overriding the last block value", async () => {
       manager.addConfigFile({
         type: "file",
         fullPath: "/dev/path/readme.md",