+++ packages/libs/configuration/src/configuration-manager/configuration-manager.ts
@@ -1,23 +1,12 @@
 import { evaluateGuard, mergeOverwriteOrAppend } from "@autorest/common";
 import { IFileSystem } from "@azure-tools/datastore";
 import { AutorestConfiguration, createAutorestConfiguration } from "../autorest-configuration";
-import { AutorestRawConfiguration } from "../autorest-raw-configuration";
+import { AutorestNormalizedConfiguration } from "../autorest-normalized-configuration";
+import { AUTOREST_INITIAL_CONFIG } from "../configuration-schema";
 import { desugarRawConfig } from "../desugar";
 import { ConditionalConfiguration, ConfigurationFile } from "./configuration-file";
 
-const initialConfig: AutorestRawConfiguration = Object.freeze({
-  "directive": [],
-  "input-file": [],
-  "exclude-file": [],
-  "profile": [],
-  "output-artifact": [],
-  "require": [],
-  "try-require": [],
-  "use": [],
-  "pass-thru": [],
-});
-
-const defaultConfig: AutorestRawConfiguration = Object.freeze({
+const defaultConfig: AutorestNormalizedConfiguration = Object.freeze({
   "base-folder": ".",
   "output-folder": "generated",
   "debug": false,
@@ -27,7 +16,7 @@ const defaultConfig: AutorestRawConfiguration = Object.freeze({
 
 interface SimpleConfiguration {
   type: "simple";
-  config: AutorestRawConfiguration;
+  config: AutorestNormalizedConfiguration;
 }
 
 type ConfigurationItem = ConfigurationFile | SimpleConfiguration;
@@ -43,7 +32,7 @@ export class ConfigurationManager {
 
   public constructor(private configFileOrFolderUri: string, private fileSystem: IFileSystem) {}
 
-  public async addConfig(config: AutorestRawConfiguration) {
+  public async addConfig(config: AutorestNormalizedConfiguration) {
     this.configItems.push({ type: "simple", config: await desugarRawConfig(config) });
   }
 
@@ -52,7 +41,7 @@ export class ConfigurationManager {
    * This means this configuration will be loaded first, its value will be able to be used in later configurations.
    * @param config Configuration.
    */
-  public async addHighPriorityConfig(config: AutorestRawConfiguration) {
+  public async addHighPriorityConfig(config: AutorestNormalizedConfiguration) {
     this.configItems.unshift({ type: "simple", config: await desugarRawConfig(config) });
   }
 
@@ -69,7 +58,7 @@ export class ConfigurationManager {
    * It will resolve potential condition for configuration file blocks to be included.
    */
   public async resolveConfig(): Promise<AutorestConfiguration> {
-    let current = initialConfig;
+    let current = AUTOREST_INITIAL_CONFIG;
 
     const configFileNames = [];
     for (const configItem of this.configItems) {
@@ -92,8 +81,11 @@ export class ConfigurationManager {
    * @param config Current config. Will be used to resolve values in the config file.(Such as condition or interpolate values).
    * @param configFile Config file.
    */
-  private mergeConfigFile(config: AutorestRawConfiguration, configFile: ConfigurationFile): AutorestRawConfiguration {
-    let currentFileResolution = { ...initialConfig };
+  private mergeConfigFile(
+    config: AutorestNormalizedConfiguration,
+    configFile: ConfigurationFile,
+  ): AutorestNormalizedConfiguration {
+    let currentFileResolution = { ...AUTOREST_INITIAL_CONFIG };
     const resolveConfig = () => mergeOverwriteOrAppend(config, currentFileResolution);
 
     for (const configBlock of configFile.configs) {
@@ -117,7 +109,7 @@ export class ConfigurationManager {
 }
 
 const shouldIncludeConditionalConfig = (
-  context: AutorestRawConfiguration,
+  context: AutorestNormalizedConfiguration,
   config: ConditionalConfiguration,
   forceAllVersionsMode: boolean,
 ) => (config.condition ? evaluateGuard(config.condition, context, forceAllVersionsMode) : true);